Tools: impove collision check script
It was kind of crappy before because I rewrote part of it halfway through
but never really finished the job. This simplifies operation a lot and increases
execution speed. Also, it now outputs collisions in a more structured (normal)
format using tabs for alignment. This will allow the output to be parsed by other
tools more easily.