stripios: Build from source
instead of putting it as a binary in the git repo, which is not
portable across different platforms (or even glibc versions on Linux)!

Note: This will break the github action for the time being
The devkitARM docker is based on debian-lite, which does not
include a C++ compiler or the apt source to install one. This
shouldn't be too big of a deal to solve, but it's not something
I have the energy for *right now*.